# Improve getGroups performance