… shell/components/nav/__tests__/Type.test.ts
… shell/store/__tests__/type-map.test.ts