<!-- This template is for Devs to give QA details before moving the issue To-Test -->
### Summary
Fixes #7773
<!-- Define findings related to the feature or bug issue. -->

> The general refactor of the side nav is covered by https://github.com/rancher/dashboard/issues/5607

### Occurred changes and/or fixed issues
- getGroups creates the entries used in the side nav (and the explorer resource search modal)
- This previously was triggered often and on large / slow systems can run slowly, both impacting UI performance
- This PR improves this performance in two ways
  - Improve performance of allTypes fn
    - This consumed the majority of time within getGroups
    - PR reduces execution time by a half
    - It does this by 
      - calculating entries for a product given all modes requests... rather than one per product per mode
      - skip calling the costly labelFor fn unless we actually need to
  - Reduce the amount of times getGroups is called
    - don't trigger on changes to counts (which changes often). Instead show counts via the component itself
      - to do this i added a dashboard store `count` getter which is now used in a couple of places
      - note - for explorer search stick them in again, as in that context it negatively impacts performance
    - only trigger of new / removed schemas, instead of any changes to schemas
      - this is more costly to calc up front, but saves an even more costly getGroups cal
    - queue up getGroup calls for namespace and namespaceMode changes... instead of executing immediately
      - on some change of mode the namespace list changes also, this results in x2 execution. i think we might even be able to remove changes on ns mode change... but chickened out in this PR
    - Fixed a bug where `updateNamespaces` used the wrong getters... which meant namespace cache was set to an empty object and then set to the correct value (which caused x2 calls to getGroups)   
- The PR also
  - Adds lots of comments
  - Adds unit tests for the new allTypes fn
  - set named functions for allTypes and getTree so they show up in performance logs
  - removed a watch on `product`, which didn't exist. there is a watch on `productId`

### Technical notes summary
- Added unit tests for allTypes
- Ensured `diff` between previous and new allTypes output matched. The only diff is the counts stuff which has now been left out


#### Tested at scale (7.5k schemas, 2k namespaces)
- Different global products (user management, fleet, cluster management, harvester manager)
- Different explorer products (cis benchmark, kubewarden, legacy)

Product - Mode | Old | New
-- | -- | --
Explorer - basic | 169.247314453125 ms
Explorer - favorite | 162.413330078125 ms
Explorer - used | 114.866943359375 ms
Explorer - all modes | | 150.678955078125 ms
Apps - basic | 111.8642578125 ms | 63.146728515625 ms
CIS - basic | 147.37890625 ms | 62.5 ms
Legacy - basic  | 109.410888671875 ms | 60.7373046875 ms
Kubewarden - basic | 115.473876953125 ms | 61.5009765625 ms
TOTAL (getProductsGroups) |  1042.572998046875 ms | 509.0517578125 ms

#### Observations
- Explorer product requests three modes
  - BASIC
    - returns configured types (probably as per product definition)
  - FAVORITE
    - user favs
  - USED
    - types that have resources. if a resource has no count at all, it won't be in this list
- Non-explorer products just request the BASIC mode
- allTypes always used with getTree and vice-versa. These might have been split at some point? Used in extension?
- labelFor getter performs much worse when passed number instead of random & invalid `counts` object.
- labelFor with a cold cache is expensive (need to cache strings... create and cache regex)
- doesn't scale with more products (though impact is now limited

### Areas or cases that should be tested
- Side nav content for each burger menu item, including a cluster. Check
  - same menu items as old
  - counts are correct, given namespace filter
  - inline products such as CIS benchmark, kubewarden, legacy  
  - favourites
  - custom nav entries
- Side nav updates given
  - new/removed resources (made via another browser) changes resource count
  - namespace/project filter changes (made in same browser) changes resource counts
- Cluster explorer search resource modal
  - same items as old
  - counts show correctly