@nwmac 2024-01-26
@codyrancher Can you review this too, please.

I see this as a tactical improvement on what we have... that nav code is clunky and needs a re-think/re-write.