@richard-cox 2024-02-05
> Favorites aren't updated between tabs for both adding and removing.

I couldn't reproduce this in master. Favourites come from  `/v1/userpreference` which isn't something we can subscribe for changes to, so this makes sense. 

> Count update is still slow, I'm not positive that it's a regression. The update time varied on the existing version but I perceived the average as being faster.

Tricky one. I had another look. Both old and new come from `count` `resource.change` socket messages. Sometimes the time varies for that to be received. In both cases, once received, the update in side nav was quick. This could be system/env specific though (if you've got high cpu on either upstream or downstream cluster). Would you be able to take another look and see if the time from `resource.change` received and the change in sidenav is laggy? 

> It looks like the navigation of at least cluster extensions had their labels changed and counts added.

KubeWarden gaining count and changing to plural... i'm kind of ok with this. If one happened without the other, it would be weirder. The pluralised names align with other menu entries and the count aligns with the number of resources ok. @jordojordo would you be able to confirm it's ok your side?

Elemental not having any counts... i'd like to investigate this further. was this a regression? looking at https://github.com/rancher/elemental-ui/blob/main/pkg/elemental/elemental-config.js it looks like they're geniune resources (not virtual or spoofed) but do have `customRoute`s. wondering if i broke that. having trouble validating at the moment (element-crd chart doesn't seem to exist)

Users & Auth Section... they're all virtual or spoofed and therefor don't have counts