@codyrancher 2024-02-06
> Would you be able to take another look and see if the time from resource.change received and the change in sidenav is laggy?

It does appear to be quick after we get the message.

> KubeWarden gaining count and changing to plural... i'm kind of ok with this.

I'm okay with it too. It just seems like it should be notated somewhere since it affects extensions.